requirements.txt
discord.py
yt-dlp
pynacl